+++ src/kernel_cheetah_decode.h
@@ -75,7 +75,7 @@ typedef struct {
     uint_fast32_t signaturesCount;
     uint_fast8_t efficiencyChecked;
 
-    uint_fast64_t endDataOverhead;
+    uint_fast8_t endDataOverhead;
 
     uint_fast16_t lastHash;
 
@@ -83,7 +83,7 @@ typedef struct {
 } density_cheetah_decode_state;
 #pragma pack(pop)
 
-DENSITY_KERNEL_DECODE_STATE density_cheetah_decode_init(density_cheetah_decode_state *, const density_main_header_parameters parameters, const uint_fast32_t);
+DENSITY_KERNEL_DECODE_STATE density_cheetah_decode_init(density_cheetah_decode_state *, const density_main_header_parameters parameters, const uint_fast8_t);
 DENSITY_KERNEL_DECODE_STATE density_cheetah_decode_continue(density_memory_teleport *, density_memory_location *, density_cheetah_decode_state *);
 DENSITY_KERNEL_DECODE_STATE density_cheetah_decode_finish(density_memory_teleport *, density_memory_location *, density_cheetah_decode_state *);
